backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
passlib==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
